requirements.txt
Flask==2.3.3
faster-whisper==1.0.1
ffmpeg-python==0.2.0
moviepy==1.0.3
srt==3.5.3
deep-translator==1.11.4
torch==2.2.2
numpy==1.26.4